requirements.txt
torch>=1.2.0
torchvision>=0.2.1
h5py>=2.7.1
scikit-image>=0.15.0
numpy>=1.17.0
tqdm>=4.35.0
pillow>=5.3.0
scipy>=1.0.1
pandas>=0.22.0
tb-nightly
